# Conflicts:
#	build.gradle
#	src/main/java/be/ddd/api/cafe/CafeBeverageAPI.java
#	src/main/java/be/ddd/api/dto/req/MemberProfileModifyDto.java
#	src/main/java/be/ddd/api/dto/req/MemberProfileRegistrationDto.java
#	src/main/java/be/ddd/api/dto/res/BeverageSearchResultDto.java
#	src/main/java/be/ddd/api/dto/res/MemberDetailsDto.java
#	src/main/java/be/ddd/api/dto/res/MemberModifyDetailsDto.java
#	src/main/java/be/ddd/application/beverage/CafeBeverageQueryService.java
#	src/main/java/be/ddd/application/beverage/CafeBeverageQueryServiceImpl.java
#	src/main/java/be/ddd/application/beverage/CafeBeverageRepositoryImpl.java
#	src/main/java/be/ddd/application/beverage/dto/BeverageSearchDto.java
#	src/main/java/be/ddd/application/member/SugarRecommendationService.java
#	src/main/java/be/ddd/domain/entity/member/Member.java
#	src/main/java/be/ddd/domain/entity/member/MemberHealthMetric.java
#	src/main/java/be/ddd/domain/repo/CafeBeverageRepositoryCustom.java
#	src/main/java/be/ddd/domain/repo/MemberRepository.java
#	src/main/java/be/ddd/domain/repo/MemberRepositoryCustom.java
#	src/main/resources/application-prod.yml
#	src/main/resources/application.yml

+++ src/main/java/be/ddd/api/dto/req/MemberProfileModifyDto.java
@@ -5,7 +5,6 @@ import be.ddd.domain.entity.member.Gender;
 import be.ddd.domain.entity.member.SugarIntakeLevel;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
@@ -14,6 +13,6 @@ public record MemberProfileModifyDto(
         @JsonFormat(pattern = "yyyy-MM-dd") LocalDate birthDay,
         Gender gender,
         Integer heightCm,
-        BigDecimal weightKg,
+        Integer weightKg,
         ActivityRange activityRange,
         SugarIntakeLevel sugarIntakeLevel) {}
