+++ src/main/java/be/ddd/api/dto/res/MemberDetailsDto.java
@@ -5,7 +5,6 @@ import be.ddd.domain.entity.member.Gender;
 import be.ddd.domain.entity.member.Member;
 import be.ddd.domain.entity.member.SugarIntakeLevel;
 import com.fasterxml.jackson.annotation.JsonFormat;
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record MemberDetailsDto(
@@ -13,7 +12,7 @@ public record MemberDetailsDto(
         @JsonFormat(pattern = "yyyy-MM-dd") LocalDate birthDay,
         Gender gender,
         Integer heightCm,
-        BigDecimal weightKg,
+        Integer weightKg,
         ActivityRange activityRange,
         SugarIntakeLevel sugarIntakeLevel) {
